server/requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
Django==5.1.3
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
mysqlclient==2.2.6
packaging==24.2
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
uritemplate==4.1.1
urllib3==2.2.3
